my_class/pyproject.toml
[tool.poetry]
name = "my_class"
version = "0.1.0"
description = ""
authors = ["Julia <julie.panov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.6.0"
psycopg2 = "^2.8.6"
Flask-Script = "^2.0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
